requirements.txt
djangorestframework
django
gunicorn
django-heroku
requests
Pillow
django-rest-auth
django-allauth